requirements.txt
bson==0.5.10
click==8.1.3
Flask==2.2.2
Flask-PyMongo==2.3.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.5.0
pymongo==4.2.0
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.2.2
zipp==3.8.1
